fix error between 32 and 64

+++ framebuffer/display.c
@@ -1,5 +1,6 @@
 #include <unistd.h>  
-#include <stdio.h>  
+#include <stdio.h> 
+#include <stdint.h> 
 #include <stdlib.h>  
 #include <fcntl.h>  
 #include <string.h>  
@@ -12,9 +13,9 @@
 typedef struct  
 {  
     char cfType[2];//文件类型，"BM"(0x4D42)  
-    long cfSize;//文件大小（字节）  
-    long cfReserved;//保留，值为0  
-    long cfoffBits;//数据区相对于文件头的偏移量（字节）  
+    uint32_t cfSize;//文件大小（字节）  
+    uint32_t cfReserved;//保留，值为0  
+    uint32_t cfoffBits;//数据区相对于文件头的偏移量（字节）  
 }__attribute__((packed)) BITMAPFILEHEADER;  
 //__attribute__((packed))的作用是告诉编译器取消结构在编译过程中的优化对齐  
   
@@ -22,10 +23,10 @@ typedef struct
 typedef struct  
 {  
     char ciSize[4];//BITMAPFILEHEADER所占的字节数  
-    long ciWidth;//宽度  
-    long ciHeight;//高度  
+    uint32_t ciWidth;//宽度  
+    uint32_t ciHeight;//高度  
     char ciPlanes[2];//目标设备的位平面数，值为1  
-    unsigned short int ciBitCount;//每个像素的位数  
+    uint16_t ciBitCount;//每个像素的位数  
     char ciCompress[4];//压缩说明  
     char ciSizeImage[4];//用字节表示的图像大小，该数据必须是4的倍数  
     char ciXPelsPerMeter[4];//目标设备的水平像素数/米
